Changed the way that test tables are handled. TableDefs.pm now has a 'debug_line' function that outputs debugging info in the proper way depending on the environment in which it was invoked, including from test files.
Added TestTables.pm, which exports 'establish_test_tables' and
'fill_test_table'. The former copies the schemas for each table
in a specified table group to the test database, and the latter
fills a specified table in the test database with rows from the
corresponding table in the main database filtered by an optional
expression.