Modified EditTester.pm so that the name of a subclass can be specified when instantiating the EditTester class. This allows for calling methods to establish the necessary tables and to complete them by e.g. adding the necessary triggers.
The method 'establish_test_tables' now calls the corresponding method
of the subclass if one exists, and otherwise calls the corresponding
routine from TestTables.pm with the appropriate arguments.

The method 'fill_test_table' was added, which calls the corresponding
routine from TestTables.pm with the appropriate arguments.

The method 'complete_test_tables' was added, which calls the corresponding
method of the subclass if one exists, and prints a warning otherwise.