All of the edt-*.t tests now create an EditTester object using the subclass 'EditTest'.
In edt-40-bulk.t, all of the subtests check for the environment
variable PBDB_TEST_NO_BULK, and if it has a true value then they
simply pass a single test and exit. This allows for a quick run
through of the main EditTransaction tests while skipping the lengthy
bulk operation tests.