Moved 'build_tables.pl' script into this repo, and rewrote it a bit. The script now takes a --log option, specifying the file to which log messages are written. Anything written to STDOUT goes to the log file. Exceptions are written both to the log and to STDERR.
It also takes the option --nightly, which is equivalent to -cmty.
The available single-letter options have been reduced to those four.

In ConsoleLog.pm, timestamps are now suffixed with the time zone.

In TaxonTables.pm, the main query statement in PopulateOpinionCache
is now written to a local variable before execution, to make
debugging easier.